Remove server dependance on SFML2.
The only thing the server used from the sfml2 lib was the Vector2f type - a very simple type.  I reimplemented this type as FloatVector in the shared package and use it everywhere.  The client has functions to convert FloatVector<->Vector2f back and forth so it can continue to use the SFML2 version which are needed in the sfml2 calls.